Use the new API for copyFile().
